split for random functions
