+++ src/bluesim/dollar_time.cxx
@@ -11,16 +11,3 @@ tUInt32 dollar_stime(tSimStateHdl simHdl)
 {
   return (tUInt32) bk_now(simHdl);
 }
-
-tUInt32 dollar_random(tSimStateHdl simHdl)
-{
-  return (tUInt32) random();
-}
-
-tUInt32 dollar_urandom_range(tSimStateHdl simHdl,
-    const char* /* size_str */,
-    tUInt32 maxval,
-    tUInt32 minval = 0)
-{
-  return (dollar_random(simHdl) % (maxval - minval + 1)) + minval;
-}
